Finish basic unit class and create spaceShip class
