+++ src/main/java/group22/space_invaders/unit/SpaceShip.java
@@ -1,0 +1,15 @@
+package group22.space_invaders.unit;
+
+/**
+ * A SpaceShip in the game extends Unit.
+ * 
+ * @author Bryan van Wijk
+ */
+
+public class SpaceShip extends Unit {
+
+	public SpaceShip(float X, float Y) {
+		super(X, Y);
+	}
+
+}
